README: Fix mounting directory for Docker run
According to the Filesystem Hierarchy Standard (FHS), the `/usr`
directory is for shareable and read-only data. A more appropriate place
to mount the working directory should actually be `/tmp`.

This commit fixes the instruction for mounting the working directory
when using `docker run` to mount in the container's `/tmp` directory
instead.

Refer to issue #21.